refactor(agents): move MCP tool-approval handling into ManagedRuntime
Tool approval is a managed-only concern, but ingress handleAction was parsing
mcp-approval ids and calling ConfirmToolApproval directly. Route it through the
runtime instead:

- Add resolved subscriberId to ConversationTurn so ManagedRuntime can build the
  ConfirmToolApprovalCommand with the exact same id ingress used (behavior-preserving).
- ManagedRuntime.dispatch now handles ON_ACTION turns: mcp-approval -> ConfirmToolApproval,
  anything else is a no-op (managed has no bridge onAction).
- Ingress no longer imports ConfirmToolApproval/parseToolApprovalActionId; it just
  builds the turn and dispatches. Link buttons stay a generic swallow in ingress.
- Relocate the generic isLinkButtonActionId helper to shared/util.

Behavior preserved (137 passing / 2 pre-existing failures).
